// frontend/src/api/model/githubCom4H1RZooraInternalDomainChangelogEntry.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */

export interface GithubCom4H1RZooraInternalDomainChangelogEntry {
  body_en?: string;
  body_fa?: string;
  created_at?: string;
  id?: string;
  is_major?: boolean;
  published_at?: string;
  title_en?: string;
  title_fa?: string;
  updated_at?: string;
  version?: string;
}